Refactor repositories and add DTOs for improved structure.
Changed ID types in repositories from Long to Integer for consistency. Introduced multiple DTOs (e.g., SprzedazDTO, GatunekDTO) to streamline data transfer objects. Implemented new service methods and updated SprzedazController to utilize DTOs, enhancing API handling and response structure.

+++ src/main/java/czechowski/treeswebappbackend/repository/GatunekRepository.java
@@ -4,5 +4,5 @@ package czechowski.treeswebappbackend.repository;
 import czechowski.treeswebappbackend.model.Gatunek;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GatunekRepository extends JpaRepository<Gatunek, Long> {
+public interface GatunekRepository extends JpaRepository<Gatunek, Integer> {
 }
