+++ src/main/java/czechowski/treeswebappbackend/repository/StoiskoRepository.java
@@ -4,5 +4,5 @@ package czechowski.treeswebappbackend.repository;
 import czechowski.treeswebappbackend.model.Stoisko;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StoiskoRepository extends JpaRepository<Stoisko, Long> {
+public interface StoiskoRepository extends JpaRepository<Stoisko, Integer> {
 }
