+++ src/main/java/czechowski/treeswebappbackend/repository/WielkoscRepository.java
@@ -4,5 +4,5 @@ package czechowski.treeswebappbackend.repository;
 import czechowski.treeswebappbackend.model.Wielkosc;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WielkoscRepository extends JpaRepository<Wielkosc, Long> {
+public interface WielkoscRepository extends JpaRepository<Wielkosc, Integer> {
 }
